requirements.txt
appnope==0.1.4
argcomplete==3.6.2
asttokens==3.0.0
attrs==25.3.0
awscli==1.38.36
backcall==0.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
boto3==1.37.36
botocore==1.37.36
caldav==1.4.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
configparser==7.2.0
cryptography==45.0.5
DateTime==5.5
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
docutils==0.19
executing==2.2.0
fastjsonschema==2.21.1
icalendar==6.1.0
idna==3.6
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==1.26.3
packaging==24.2
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
